refactored webpack config into build-config dir using webpack-merge
