+++ build-config/fragments/common.js
@@ -1,0 +1,38 @@
+/* eslint-env node */
+
+const path = require('path');
+const webpack = require('webpack');
+
+module.exports = {
+    context: path.resolve(__dirname, '../', '../'),
+    output: {
+        libraryTarget: 'umd',
+        umdNamedDefine: true
+    },
+    performance: {
+        hints: false
+    },
+    module: {
+        rules: [
+            {
+                test: /\.js$/,
+                enforce: 'pre',
+                exclude: /node_modules/,
+                use: [
+                    {
+                        loader: 'eslint-loader'
+                    }
+                ]
+            },
+            {
+                test: /\.js$/,
+                exclude: /node_modules/,
+                use: [
+                    {
+                        loader: 'babel-loader'
+                    }
+                ]
+            }
+        ]
+    }
+};
