+++ build-config/fragments/min.js
@@ -1,0 +1,12 @@
+/* eslint-env node */
+
+const path = require('path');
+const webpack = require('webpack');
+
+module.exports = {
+    plugins: [
+        new webpack.optimize.UglifyJsPlugin({
+            sourceMap: true
+        })
+    ]
+};
