+++ build-config/fragments/prod.js
@@ -1,0 +1,5 @@
+/* eslint-env node */
+
+module.exports = {
+    devtool: 'source-map'
+};
