+++ build-config/webpack.dev.main.js
@@ -1,0 +1,8 @@
+const merge = require('webpack-merge');
+const path = require('path');
+
+const common = require('./fragments/common');
+const dev = require('./fragments/dev');
+const main = require('./fragments/main');
+
+module.exports = merge(common, dev, main);
