+++ build-config/webpack.prod.htmlinit.js
@@ -1,0 +1,8 @@
+const merge = require('webpack-merge');
+const path = require('path');
+
+const common = require('./fragments/common');
+const prod = require('./fragments/prod');
+const htmlinit = require('./fragments/htmlinit');
+
+module.exports = merge(common, prod, htmlinit);
