+++ build-config/webpack.prod.plugins.js
@@ -1,0 +1,8 @@
+const merge = require('webpack-merge');
+const path = require('path');
+
+const common = require('./fragments/common');
+const prod = require('./fragments/prod');
+const plugins = require('./fragments/plugins');
+
+module.exports = merge(common, prod, plugins);
