+++ package.json
@@ -7,13 +7,19 @@
     "example": "example"
   },
   "scripts": {
-    "start": "webpack --env.plugins && webpack --env.htmlinit && webpack-dev-server",
-    "start:htmlinit": "webpack && webpack --env.plugins && webpack-dev-server --env.htmlinit",
-    "start:plugins": "webpack && webpack --env.htmlinit && webpack-dev-server --env.plugins",
+    "start": "npm run build:plugins && npm run build:htmlinit && webpack-dev-server --config ./build-config/webpack.dev.main.js",
+    "start:htmlinit": "npm run build:main && npm run build:plugins && webpack-dev-server --config ./build-config/webpack.dev.htmlinit.js",
+    "start:plugins": "npm run build:main && npm run build:htmlinit && webpack-dev-server --config ./build-config/webpack.dev.plugins.js",
     "build": "npm run build:normal && npm run build:minified",
-    "build:normal": "webpack && webpack --env.plugins && webpack --env.htmlinit",
-    "build:minified": "webpack --env.minify && webpack --env.plugins --env.minify && webpack --env.htmlinit --env.minify",
     "format": "prettier --single-quote true --tab-width 4 'src/**/*.js*' 'example/**/*.js*' 'spec/**/*.js*' 'website/**/*.js*' --write",
+    "build:main": "webpack --config ./build-config/webpack.prod.main.js",
+    "build:main:min": "webpack --config ./build-config/webpack.prod.main.min.js",
+    "build:plugins": "webpack --config ./build-config/webpack.prod.plugins.js",
+    "build:plugins:min": "webpack --config ./build-config/webpack.prod.plugins.min.js",
+    "build:htmlinit": "webpack --config ./build-config/webpack.prod.htmlinit.js",
+    "build:htmlinit:min": "webpack --config ./build-config/webpack.prod.htmlinit.min.js",
+    "build:normal": "npm run build:main && npm run build:plugins && npm run build:htmlinit",
+    "build:minified": "npm run build:main:min && npm run build:plugins:min && npm run build:htmlinit:min",
     "lint": "eslint src example spec website",
     "lint-and-format-staged": "lint-staged",
     "doc": "esdoc",
@@ -58,6 +64,7 @@
     "prettier": "^1.7.3",
     "webpack": "3",
     "webpack-dev-server": "2"
+    "webpack-merge": "^4.1.1"
   },
   "homepage": "https://github.com/katspaugh/wavesurfer.js",
   "pre-commit": "lint-and-format-staged",
